python_app/requirements.txt
kubernetes
google-cloud-storage
google.cloud
